rtwn-usbpcap: add a completion cache, populate it, not using it yet
Prototype up a completion cache so i can match up the submission and
completion per endpoint.  It's stupid and O(n) for now but that's
fine for what I'm doing.

I'm not using it yet; it runs but doens't crash.  Next up, rewriting
the decoders to care about the submission/completion bits and when
we only get one or the other (eg for errors.)